# backend/swagger.yaml
openapi: '3.0.3'
info:
  title: Self host event managemt
  description: Made by Harpreet Singh At @redergo
  version: 1.1.0
  contact:
    email: harpreet.singh@redergo.com, h.singh8149@gmail.com
    url: 'https://redergo.com/en/'
servers:
  - url: '{protocol}://{url}/api/{version}'
    description: for local host
    variables:
      version:
        enum:
          - v1
        default: v1
      url:
        enum:
          - localhost:3222
          - api-dev-shem.hsingh.site
        default: localhost:3222
      protocol:
        enum:
          - http
          - https
        default: http
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /user/refresh-token:
    
    post:
      summary: Generate a new refresh token
      description: Generate a new refresh token for a user with a valid access token
      operationId: generateRefreshToken
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                  description: The user's valid access token
              required:
                - accessToken
      responses:
        '200':
          description: Refresh token generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  refreshToken:
                    type: string
        '400':
          description: Access token is required or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Error in generating refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
...
  /users/logout-all-devices:
    get:
      tags:
        - User
      summary: Log out a user from all devices
      description: Log out a user from all devices using a valid access token
      operationId: logoutAllDevices
      responses:
        '200':
          description: User logged out successfully from all devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized, valid access token required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Error in logging out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
  /users/password-reset/request-jwt:
    post:
      tags:
        - User
      summary: Generate a JWT for password reset
      description: Generate a JWT for password reset
      operationId: forgotPasswordJwt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: The user's email address
              required:
                - email
      responses:
        '200':
          description: JWT for password reset generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jwt:
                    type: string
        '400':
          description: Email is required or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Error in generating JWT for password reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
  /users/password-reset/confirm-jwt/:token:
    post:
      tags:
        - User
      summary: Reset password using JWT token
      description: Reset password using JWT token
      operationId: passwordResetJwtToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jwt:
                  type: string
                  description: The JWT token for password reset
                newPassword:
                  type: string
                  description: The new password
              required:
                - jwt
                - newPassword
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: JWT and new password are required or invalid JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Error in resetting password
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
  /users/profile:
    get:
      tags:
        - User
      summary: Get the details of the logged in user
      description: Get the details of the logged in user
      operationId: getLoggedInUserDetails
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    type: object
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                      surname:
                        type: string
                      phoneNumber:
                        type: string
                      deviceId:
                        type: string
        '401':
          description: Unauthorized, valid access token required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Error in retrieving user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string